Address minor tech debts in main test suite
